requirements-retail.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.38.0
boto3==1.35.24
pandas==2.2.2
pyarrow==17.0.0
kafka-python==2.0.2
pyspark==3.5.1
